no more globals (sorta)
