+++ MW_OSD/globals.h
@@ -19,7 +19,8 @@ struct Timer_s {
   uint32_t seconds;
   uint8_t MSP_active;
   uint8_t GPS_active;
-  uint8_t armed;
+  // armed timer
+  uint8_t armed = 255;
   uint32_t allSec;
 };
 
@@ -48,130 +49,63 @@ struct Mode_s {
   uint32_t gpsland;
 };
 
+// RC Info
+struct RC_s {
+  uint8_t rcRate8,rcExpo8;
+  uint8_t rollPitchRate;
+  uint8_t rollRate;
+  uint8_t PitchRate;
+  uint8_t yawRate;
+  uint8_t dynThrPID;
+  uint8_t thrMid8;
+  uint8_t thrExpo8;
+  uint8_t tpa_breakpoint16;
+  uint8_t rcYawExpo8;
+
+  // uint8_t ?
+  int16_t rssi =0;
+  // uint8_t ?
+  int16_t oldrssi;
+  int16_t pwmRSSI = 0;
+  uint16_t MwRssi=0;
+
+  //For Current Throttle
+  uint16_t LowT = LOWTHROTTLE;
+  uint16_t HighT = HIGHTHROTTLE;
+};
+
+// MWOSD's current state
+//  stuff here should pry be moved, its mostly here b/c idk where else to put it
+struct Mwosd_s {
+  // kinda a misnomer, because the osd never really "arms", but it does track armed state independently of the FC
+  bool armed;
+  bool previousarmedstatus;
+  uint16_t pMeterSum=0;
+
+  // For Settings Defaults
+  uint8_t P8[PIDITEMS];
+  uint8_t I8[PIDITEMS];
+  uint8_t D8[PIDITEMS];
+
+  // temperature in degrees Centigrade
+  int16_t temperature=0;
+
+  // This hold barometric value
+  int32_t  old_MwAltitude=0;
+};
+
+struct Debug_s {
+  //uint16_t screen_buffer[4];
+};
+
 extern struct Timer_s timer;
 extern struct Flags_s flags;
 extern struct Mode_s mode;
+extern struct RC_s rc;
+extern struct Mwosd_s mwosd;
+extern struct Debug_s debug;
 
-//extern uint16_t debug[4];   // int32_t ?...
-//extern unsigned int allSec;
-//extern unsigned int menuSec;
-//extern uint16_t debugerror;
-//extern uint16_t debugval;
-//extern uint16_t cell_data[6];
-//extern uint16_t cycleTime;
-//extern uint16_t I2CError;
-//extern uint8_t oldROW;
-//extern uint8_t cells;
-//
-//#if defined CORRECT_MSP_BF1
-//extern uint8_t bfconfig[25];
-//#endif
-//
-//
-//// Config status and cursor location
-//extern uint8_t screenlayout;
-//extern uint8_t oldscreenlayout;
-//extern uint8_t ROW;
-//extern uint8_t COL;
-//extern int8_t configPage;
-//extern int8_t previousconfigPage;
-//extern uint8_t configMode;
-//extern uint8_t fontData[54];
-//extern uint8_t nextCharToRequest;
-//extern uint8_t lastCharToRequest;
-//
-//// For Settings Defaults
-//extern uint8_t P8[PIDITEMS], I8[PIDITEMS], D8[PIDITEMS];
-//
-//extern uint8_t rcRate8,rcExpo8;
-//extern uint8_t rollPitchRate;
-//extern uint8_t rollRate;
-//extern uint8_t PitchRate;
-//extern uint8_t yawRate;
-//extern uint8_t dynThrPID;
-//extern uint8_t thrMid8;
-//extern uint8_t thrExpo8;
-//extern uint8_t tpa_breakpoint16;
-//extern uint8_t rcYawExpo8;
-//extern uint8_t FCProfile;
-//extern uint8_t PreviousFCProfile;
-//extern uint8_t CurrentFCProfile;
-//extern uint8_t PIDController;
-//extern uint16_t LoopTime;
-//
-//extern int32_t  MwAltitude;                         // This hold barometric value
-//extern int32_t  old_MwAltitude;                     // This hold barometric value
-//
-//
-//extern int MwAngle[2];           // Those will hold Accelerator Angle
-//extern uint16_t MwRcData[8];
-//
-//// for analogue / PWM sensor filtering
-//extern int16_t sensorfilter[SENSORTOTAL][SENSORFILTERSIZE+2];
-//
-//extern uint8_t MwVersion;
-//extern uint8_t MwVBat;
-//extern int16_t MwVario;
-//extern uint8_t armed;
-//extern uint8_t previousarmedstatus;  // for statistics after disarming
-//extern uint16_t armedangle;           // for capturing direction at arming
-//extern uint32_t GPS_distanceToHome;
-//extern uint8_t GPS_fix;
-//extern uint8_t GPS_frame_timer;
-//extern int32_t GPS_latitude;
-//extern int32_t GPS_longitude;
-//extern int16_t GPS_altitude;
-//extern int16_t GPS_home_altitude;
-//extern int16_t previousfwaltitude;
-//extern int16_t interimfwaltitude;
-//extern uint16_t GPS_speed;
-//extern int16_t  GPS_ground_course;
-//extern uint16_t old_GPS_speed;
-//extern int16_t GPS_directionToHome;
-//extern uint8_t GPS_numSat;
-//extern uint8_t GPS_waypoint_step;
-////uint16_t I2CError;
-////uint16_t cycleTime;
-//extern uint16_t pMeterSum;
-//extern uint16_t MwRssi;
-//extern uint32_t GPS_time ;        //local time of coord calc - haydent
-//
-//#ifdef HAS_ALARMS
-//extern uint8_t alarmState ;
-//extern uint8_t alarmMsg[MAX_ALARM_LEN];
-//#endif
-//
-//extern uint8_t MvVBatMinCellVoltage;
-//extern uint8_t MvVBatMaxCellVoltage;
-//extern uint8_t MvVBatWarningCellVoltage;
-//
-////For Current Throttle
-//extern uint16_t LowT ;
-//extern uint16_t HighT ;
-//
-//// For Heading
-//extern int16_t MwHeading;
-//
-//// For Amperage
-//extern float amperage ;                // its the real value x10
-//extern float amperagesum ;
-//extern uint16_t MWAmperage;
-//
-//// Rssi
-//extern int16_t rssi ;   // uint8_t ?
-//extern int16_t oldrssi;   // uint8_t ?
-//extern int16_t pwmRSSI ;
-////int rssiADC;
-////int rssi_Int;
-//
-//
-//// For Voltage
-//extern uint16_t voltage;                      // its the value x10
-//extern uint16_t vidvoltage;                   // its the value x10
-//extern uint8_t voltageWarning;
-//extern uint8_t vidvoltageWarning;
-//
-//// For temprature
-//extern int16_t temperature;                  // temperature in degrees Centigrade
+extern uint8_t Settings[EEPROM_SETTINGS];
+extern uint16_t Settings16[EEPROM16_SETTINGS];
 
 #endif
