+++ MW_OSD/msp.cpp
@@ -75,7 +75,7 @@ void MSPClass::BuildRequests() {
   case REQ_MSP_ATTITUDE:
     _MSPcmdsend = MSP_ATTITUDE;
     break;
-#endif //MSP_SPEED_LOW  
+#endif //MSP_SPEED_LOW
   case REQ_MSP_ALTITUDE:
     _MSPcmdsend = MSP_ALTITUDE;
     break;
@@ -96,7 +96,7 @@ void MSPClass::BuildRequests() {
     break;
   case REQ_MSP_LOOP_TIME:
     _MSPcmdsend = MSP_LOOP_TIME;
-    break;        
+    break;
   case REQ_MSP_BOX:
 #ifdef BOXNAMES
     _MSPcmdsend = MSP_BOXNAMES;
@@ -137,7 +137,7 @@ void MSPClass::BuildRequests() {
 }
 
 void MSPClass::SendRequests() {
-  WriteRequest(_MSPcmdsend, 0);      
+  WriteRequest(_MSPcmdsend, 0);
 };
 
 void MSPClass::SetRequests() {
@@ -145,7 +145,7 @@ void MSPClass::SetRequests() {
     _modeMSPRequests = REQ_MSP_FONT;
   }
   else if(configMode) {
-    _modeMSPRequests = 
+    _modeMSPRequests =
       REQ_MSP_IDENT|
       REQ_MSP_STATUS|
       REQ_MSP_RAW_GPS|
@@ -171,26 +171,26 @@ void MSPClass::SetRequests() {
       REQ_MSP_RC;
   }
   else {
-    _modeMSPRequests = 
+    _modeMSPRequests =
       REQ_MSP_STATUS|
       REQ_MSP_RC|
      #ifdef DEBUGMW
       REQ_MSP_DEBUG|
      #endif
-     #ifdef SPORT      
+     #ifdef SPORT
       REQ_MSP_CELLS|
      #endif
      #ifdef HAS_ALARMS
       REQ_MSP_ALARMS|
      #endif
       REQ_MSP_ATTITUDE;
-    if(Sensors.IsPresent(BAROMETER)){ 
+    if(Sensors.IsPresent(BAROMETER)){
       _modeMSPRequests |= REQ_MSP_ALTITUDE;
     }
     if(flags.ident!=1){
       _modeMSPRequests |= REQ_MSP_IDENT;
     }
-    if(Sensors.IsPresent(GPSSENSOR)) 
+    if(Sensors.IsPresent(GPSSENSOR))
       _modeMSPRequests |= REQ_MSP_RAW_GPS| REQ_MSP_COMP_GPS;
     if(mode.armed == 0)
       _modeMSPRequests |=REQ_MSP_BOX;
@@ -199,11 +199,11 @@ void MSPClass::SetRequests() {
     _modeMSPRequests |= REQ_MSP_NAV_STATUS;
 #endif
   }
- 
+
   if(Settings[S_MAINVOLTAGE_VBAT] ||
     Settings[S_MWRSSI]) {
     _modeMSPRequests |= REQ_MSP_ANALOG;
-    
+
 #ifdef USE_FC_VOLTS_CONFIG
     _modeMSPRequests |= REQ_MSP_MISC;
 #endif
@@ -259,7 +259,7 @@ void MSPClass::serialMSPCheck()
       write8(OSD_SENSORS);
       for (uint8_t sensor=0;sensor<SENSORTOTAL;sensor++) {
 //        uint16_t sensortemp = analogRead(sensorpinarray[sensor]);
-        uint16_t sensortemp = (uint16_t)sensorfilter[sensor][SENSORFILTERSIZE]/SENSORFILTERSIZE;
+        uint16_t sensortemp = (uint16_t)Sensors.sensorfilter[sensor][SENSORFILTERSIZE]/SENSORFILTERSIZE;
         write16(sensortemp);
       }
        writeChecksum();
@@ -276,24 +276,24 @@ void MSPClass::serialMSPCheck()
       }
       else if(_dataSize == 56) {
         for(uint8_t i = 0; i < 54; i++)
-          fontData[i] = read8();
-      
+          Max.fontData[i] = read8();
+
 	uint8_t c = read8();
-        MAX7456.WriteNvm(c);
+        Max.WriteNvm(c);
 	//fontCharacterReceived(c);
         if (c==255)
-          MAX7456.Setup();
+          Max.Setup();
       }
     }
     if(cmd == OSD_DEFAULT) {
-      Eeprom.ClearSettings(); 
+      Eeprom.ClearSettings();
       Eeprom.CheckSettings();
       flags.reset=1;
     }
     if(cmd == OSD_RESET) {
         flags.reset=1;
     }
-                    
+
   }
 
 #ifndef GPSOSD
@@ -311,8 +311,8 @@ void MSPClass::serialMSPCheck()
     Sensors.SetActive(read32());
     #if defined FORCESENSORS
       Sensors.Force();
-    #endif  
-    armed = (Sensors.IsActive(mode.armed)) != 0;
+    #endif
+    mwosd.armed = (Sensors.IsActive(mode.armed)) != 0;
     FCProfile = read8();
     if (!configMode){
       CurrentFCProfile=FCProfile;
@@ -332,40 +332,40 @@ void MSPClass::serialMSPCheck()
     #ifdef GPSACTIVECHECK
      timer.GPS_active=GPSACTIVECHECK;
     #endif //GPSACTIVECHECK
-    uint8_t GPS_fix_temp=read8();
-    if (GPS_fix_temp){
-      GPS_fix=1;
+    uint8_t Gps_fix_temp=read8();
+    if (Gps_fix_temp){
+      Gps.fix=1;
     }
-    GPS_numSat=read8();
-    GPS_latitude = read32();
-    GPS_longitude = read32();
-    GPS_altitude = read16();
+    Gps.numSat = read8();
+    Gps.latitude = read32();
+    Gps.longitude = read32();
+    Gps.altitude = read16();
     #if defined RESETGPSALTITUDEATARM
-      if (!armed){
-        GPS_home_altitude=GPS_altitude;
-      } 
-      GPS_altitude=GPS_altitude-GPS_home_altitude;
-    #endif // RESETGPSALTITUDEATARM  
+      if (!mwosd.armed){
+        Gps.home_altitude=Gps.altitude;
+      }
+      Gps.altitude=Gps.altitude-Gps.home_altitude;
+    #endif // RESETGPSALTITUDEATARM
     #if defined I2CGPS_SPEED
-      GPS_speed = read16()*10;
+      Gps.speed = read16()*10;
       //gpsfix(); untested
     #else
-      GPS_speed = read16();
+      Gps.speed = read16();
     #endif // I2CGPS_SPEED
-    GPS_ground_course = read16();
+    Gps.ground_course = read16();
   }
 
   if (_cmdMSP==MSP_COMP_GPS)
   {
-    GPS_distanceToHome=read16();
+    Gps.distanceToHome=read16();
 #ifdef I2CGPS_DISTANCE
     gpsdistancefix();
 #endif
-    
-    GPS_directionToHome=read16();
+
+    Gps.directionToHome=read16();
 #ifdef GPSTIME
     read8(); //missing
-    GPS_time = read32();        //local time of coord calc - haydent
+    Gps.time = read32();        //local time of coord calc - haydent
 #endif
   }
 
@@ -375,7 +375,7 @@ void MSPClass::serialMSPCheck()
      read8();
      read8();
      read8();
-     GPS_waypoint_step=read8();
+     Gps.waypoint_step=read8();
   }
 #endif //MULTIWII_V24
 
@@ -384,12 +384,12 @@ void MSPClass::serialMSPCheck()
     for(uint8_t i=0;i<2;i++){
       MwAngle[i] = read16();
     }
-      MwHeading = read16();
+      Gps.MwHeading = read16();
     #if defined(USEGPSHEADING)
-      MwHeading = GPS_ground_course/10;
+      Gps.MwHeading = Gps.ground_course/10;
     #endif
     #ifdef HEADINGCORRECT
-      if (MwHeading >= 180) MwHeading -= 360;
+      if (Gps.MwHeading >= 180) Gps.MwHeading -= 360;
     #endif
   }
 
@@ -397,7 +397,7 @@ void MSPClass::serialMSPCheck()
   if (_cmdMSP==MSP_DEBUG)
   {
     for(uint8_t i=0;i<4;i++)
-      debug[i] = read16();
+      Screen.debugBuffer[i] = read16();
  }
 #endif
 #ifdef SPORT
@@ -410,10 +410,10 @@ void MSPClass::serialMSPCheck()
   if (_cmdMSP==MSP_ALTITUDE)
   {
     #if defined(USEGPSALTITUDE)
-      MwAltitude = (int32_t)GPS_altitude*100;
+      Gps.MwAltitude = (int32_t)Gps.altitude*100;
       gpsvario();
-    #else    
-      MwAltitude =read32();
+    #else
+      Gps.MwAltitude =read32();
       MwVario = read16();
     #endif
   }
@@ -421,9 +421,9 @@ void MSPClass::serialMSPCheck()
   if (_cmdMSP==MSP_ANALOG)
   {
     MwVBat=read8();
-    pMeterSum=read16();
-    MwRssi = read16();
-    MWAmperage = read16();
+    mwosd.pMeterSum=read16();
+    rc.MwRssi = read16();
+    Stats.MWAmperage = read16();
  }
 
 #ifdef USE_FC_VOLTS_CONFIG
@@ -436,7 +436,7 @@ void MSPClass::serialMSPCheck()
     read16(); //ignore: mincommand
 
     read16(); //ignore: failsafe_throttle
-    
+
     read8(); //ignore: gps_type
     read8(); //ignore: gps_baudrate
     read8(); //ignore: gps_ubx_sbas
@@ -451,55 +451,55 @@ void MSPClass::serialMSPCheck()
     MvVBatMinCellVoltage = read8(); //vbatmincellvoltage
     MvVBatMaxCellVoltage = read8(); //vbatmaxcellvoltage
     MvVBatWarningCellVoltage = read8(); //vbatwarningcellvoltage
-    
+
   }
 #endif //USE_FC_VOLTS_CONFIG
 
-#if defined (CORRECT_MSP_BF1)  
+#if defined (CORRECT_MSP_BF1)
   if (_cmdMSP==MSP_CONFIG)
   {
     for(uint8_t i=0; i<25; i++) {
-      bfconfig[i]=read8();
+      _bfconfig[i]=read8();
     }
-    rollRate = bfconfig[18];
-    PitchRate = bfconfig[19];
-    _modeMSPRequests &=~ REQ_MSP_CONFIG;    
+    rollRate = _bfconfig[18];
+    PitchRate = _bfconfig[19];
+    _modeMSPRequests &=~ REQ_MSP_CONFIG;
   }
-#endif  
-  
+#endif
+
   if (_cmdMSP==MSP_RC_TUNING)
   {
     #ifdef CORRECT_MSP_CF2
-      rcRate8 = read8();
-      rcExpo8 = read8();
-      rollRate = read8();
-      PitchRate = read8();
-      yawRate = read8();
-      dynThrPID = read8();
-      thrMid8 = read8();
-      thrExpo8 = read8();
-      tpa_breakpoint16 = read16();
-      rcYawExpo8 = read8();
+      rc.rcRate8 = read8();
+      rc.rcExpo8 = read8();
+      rc.rollRate = read8();
+      rc.PitchRate = read8();
+      rc.yawRate = read8();
+      rc.dynThrPID = read8();
+      rc.thrMid8 = read8();
+      rc.thrExpo8 = read8();
+      rc.tpa_breakpoint16 = read16();
+      rc.rcYawExpo8 = read8();
       _modeMSPRequests &=~ REQ_MSP_RC_TUNING;
     #elif defined CORRECT_MSP_CF1
-      rcRate8 = read8();
-      rcExpo8 = read8();
-      rollRate = read8();
-      PitchRate = read8();
-      yawRate = read8();
-      dynThrPID = read8();
-      thrMid8 = read8();
-      thrExpo8 = read8();
-      tpa_breakpoint16 = read16();
+      rc.rcRate8 = read8();
+      rc.rcExpo8 = read8();
+      rc.rollRate = read8();
+      rc.PitchRate = read8();
+      rc.yawRate = read8();
+      rc.dynThrPID = read8();
+      rc.thrMid8 = read8();
+      rc.thrExpo8 = read8();
+      rc.tpa_breakpoint16 = read16();
       _modeMSPRequests &=~ REQ_MSP_RC_TUNING;
     #else
-      rcRate8 = read8();
-      rcExpo8 = read8();
-      rollPitchRate = read8();
-      yawRate = read8();
-      dynThrPID = read8();
-      thrMid8 = read8();
-      thrExpo8 = read8();
+      rc.rcRate8 = read8();
+      rc.rcExpo8 = read8();
+      rc.rollPitchRate = read8();
+      rc.yawRate = read8();
+      rc.dynThrPID = read8();
+      rc.thrMid8 = read8();
+      rc.thrExpo8 = read8();
       _modeMSPRequests &=~ REQ_MSP_RC_TUNING;
     #endif
   }
@@ -507,9 +507,9 @@ void MSPClass::serialMSPCheck()
   if (_cmdMSP==MSP_PID)
   {
     for(uint8_t i=0; i<PIDITEMS; i++) {
-      P8[i] = read8();
-      I8[i] = read8();
-      D8[i] = read8();
+      mwosd.P8[i] = read8();
+      mwosd.I8[i] = read8();
+      mwosd.D8[i] = read8();
     }
     _modeMSPRequests &=~ REQ_MSP_PID;
 
@@ -601,7 +601,7 @@ void MSPClass::serialMSPCheck()
       --remaining;
     }
   }
-#else  
+#else
   if(_cmdMSP==MSP_BOXIDS) {
     flags.box=1;
     uint32_t bit = 1;
@@ -672,7 +672,7 @@ void MSPClass::serialMSPCheck()
       case 29:
         mode.acroplus |= bit;
         break;
-#endif //ACROPLUS        
+#endif //ACROPLUS
       }
       bit <<= 1;
       --remaining;
@@ -704,18 +704,18 @@ void MSPClass::Receive(uint8_t loops)
   {
     c = _serial->read();
 
-    #ifdef GPSOSD    
+    #ifdef GPSOSD
       timer.armed = 0;
       #if defined (NAZA)
         NAZA_NewData(c);
       #else
-        if (GPS_newFrame(c)) GPS_NewData();  
-      #endif //NAZA  
-    #endif //GPSOSD   
+        if (Gps.newFrame(c)) Gps.NewData();
+      #endif //NAZA
+    #endif //GPSOSD
 
     #if defined (MAVLINK)
        MAVLINK_NewData(c);
-    #endif //MAVLINK   
+    #endif //MAVLINK
 
     if (c_state == IDLE)
     {
@@ -770,7 +770,7 @@ void MSPClass::Receive(uint8_t loops)
 
 void MSPClass::serialMenuCommon()
   {
-    if((ROW==10)&&(COL==3)) {
+    if((Screen.ROW==10)&&(Screen.COL==3)) {
       if (_menudir>1){
         _menudir=1;
       }
@@ -778,136 +778,136 @@ void MSPClass::serialMenuCommon()
         _menudir=-1;
       }
 //      constrain(_menudir,-1,1);
-      configPage=configPage+_menudir;
+      Screen.configPage=Screen.configPage+_menudir;
     }
-    if(configPage<MINPAGE) configPage = MAXPAGE;
-    if(configPage>MAXPAGE) configPage = MINPAGE;
+    if(Screen.configPage<MINPAGE) Screen.configPage = MAXPAGE;
+    if(Screen.configPage>MAXPAGE) Screen.configPage = MINPAGE;
 #ifdef MENU_PID
-	if(configPage == MENU_PID) {
-	  if(ROW >= 1 && ROW <= 7) {
-            uint8_t MODROW=ROW-1;
-            if (ROW>5){
-              MODROW=ROW+1;
+	if(Screen.configPage == MENU_PID) {
+	  if(Screen.ROW >= 1 && Screen.ROW <= 7) {
+            uint8_t MODROW=Screen.ROW-1;
+            if (Screen.ROW>5){
+              MODROW=Screen.ROW+1;
             }
-  	    if(COL==1) P8[MODROW]=P8[MODROW]+_menudir;
-	    if(COL==2) I8[MODROW]=I8[MODROW]+_menudir;
-	    if(COL==3) D8[MODROW]=D8[MODROW]+_menudir;
+  	    if(Screen.COL==1) mwosd.P8[MODROW]=mwosd.P8[MODROW]+_menudir;
+	    if(Screen.COL==2) mwosd.I8[MODROW]=mwosd.I8[MODROW]+_menudir;
+	    if(Screen.COL==3) mwosd.D8[MODROW]=mwosd.D8[MODROW]+_menudir;
 	  }
 	}
 #endif
 #ifdef MENU_RC
         #if defined CORRECT_MENU_RCT2
-          if(configPage == MENU_RC && COL == 3) {
-	    if(ROW==1) rcRate8=rcRate8+_menudir;
-	    if(ROW==2) rcExpo8=rcExpo8+_menudir;
-	    if(ROW==3) rollRate=rollRate+_menudir;
-	    if(ROW==4) PitchRate=PitchRate+_menudir;
-	    if(ROW==5) yawRate=yawRate+_menudir;
-	    if(ROW==6) dynThrPID=dynThrPID+_menudir;
-	    if(ROW==7) thrMid8=thrMid8+_menudir;
-	    if(ROW==8) thrExpo8=thrExpo8+_menudir;
-	    if(ROW==9) tpa_breakpoint16=tpa_breakpoint16+_menudir;
+          if(Screen.configPage == MENU_RC && Screen.COL == 3) {
+	    if(Screen.ROW==1) rc.rcRate8=rc.rcRate8+_menudir;
+	    if(Screen.ROW==2) rc.rcExpo8=rc.rcExpo8+_menudir;
+	    if(Screen.ROW==3) rc.rollRate=rc.rollRate+_menudir;
+	    if(Screen.ROW==4) rc.PitchRate=rc.PitchRate+_menudir;
+	    if(Screen.ROW==5) rc.yawRate=rc.yawRate+_menudir;
+	    if(Screen.ROW==6) rc.dynThrPID=rc.dynThrPID+_menudir;
+	    if(Screen.ROW==7) rc.thrMid8=rc.thrMid8+_menudir;
+	    if(Screen.ROW==8) rc.thrExpo8=rc.thrExpo8+_menudir;
+	    if(Screen.ROW==9) rc.tpa_breakpoint16=rc.tpa_breakpoint16+_menudir;
           }
         #elif defined CORRECT_MENU_RCT1
-          if(configPage == MENU_RC && COL == 3) {
-	    if(ROW==1) rcRate8=rcRate8+_menudir;
-	    if(ROW==2) rcExpo8=rcExpo8+_menudir;
-	    if(ROW==3) rollRate=rollRate+_menudir;
-	    if(ROW==4) PitchRate=PitchRate+_menudir;
-	    if(ROW==5) yawRate=yawRate+_menudir;
-	    if(ROW==6) dynThrPID=dynThrPID+_menudir;
-	    if(ROW==7) thrMid8=thrMid8+_menudir;
-	    if(ROW==8) thrExpo8=thrExpo8+_menudir;
+          if(Screen.configPage == MENU_RC && Screen.COL == 3) {
+	    if(Screen.ROW==1) rc.rcRate8=rc.rcRate8+_menudir;
+	    if(Screen.ROW==2) rc.rcExpo8=rc.rcExpo8+_menudir;
+	    if(Screen.ROW==3) rc.rollRate=rc.rollRate+_menudir;
+	    if(Screen.ROW==4) rc.PitchRate=rc.PitchRate+_menudir;
+	    if(Screen.ROW==5) rc.yawRate=rc.yawRate+_menudir;
+	    if(Screen.ROW==6) rc.dynThrPID=rc.dynThrPID+_menudir;
+	    if(Screen.ROW==7) rc.thrMid8=rc.thrMid8+_menudir;
+	    if(Screen.ROW==8) rc.thrExpo8=rc.thrExpo8+_menudir;
          }
         #else
-          if(configPage == MENU_RC && COL == 3) {
-	    if(ROW==1) rcRate8=rcRate8+_menudir;
-	    if(ROW==2) rcExpo8=rcExpo8+_menudir;
-	    if(ROW==3) rollPitchRate=rollPitchRate+_menudir;
-	    if(ROW==4) yawRate=yawRate+_menudir;
-	    if(ROW==5) dynThrPID=dynThrPID+_menudir;
-	    if(ROW==6) thrMid8=thrMid8+_menudir;
-	    if(ROW==7) thrExpo8=thrExpo8+_menudir;
+          if(Screen.configPage == MENU_RC && Screen.COL == 3) {
+	    if(Screen.ROW==1) rc.rcRate8=rc.rcRate8+_menudir;
+	    if(Screen.ROW==2) rc.rcExpo8=rc.rcExpo8+_menudir;
+	    if(Screen.ROW==3) rc.rollPitchRate=rc.rollPitchRate+_menudir;
+	    if(Screen.ROW==4) rc.yawRate=rc.yawRate+_menudir;
+	    if(Screen.ROW==5) rc.dynThrPID=rc.dynThrPID+_menudir;
+	    if(Screen.ROW==6) rc.thrMid8=rc.thrMid8+_menudir;
+	    if(Screen.ROW==7) rc.thrExpo8=rc.thrExpo8+_menudir;
 	  }
         #endif
 #endif
 #ifdef MENU_VOLTAGE
-	if(configPage == MENU_VOLTAGE && COL == 3) {
-	  if(ROW==1) Settings[S_DISPLAYVOLTAGE]=!Settings[S_DISPLAYVOLTAGE];  
-	  if(ROW==2) Settings[S_DIVIDERRATIO]=Settings[S_DIVIDERRATIO]+_menudir;
-	  if(ROW==3) Settings[S_VOLTAGEMIN]=Settings[S_VOLTAGEMIN]+_menudir;
-	  if(ROW==4) Settings[S_VIDVOLTAGE]=!Settings[S_VIDVOLTAGE];
-	  if(ROW==5) Settings[S_VIDDIVIDERRATIO]=Settings[S_VIDDIVIDERRATIO]+_menudir;
-	  if(ROW==6) Settings[S_BATCELLS]=Settings[S_BATCELLS]+_menudir;
-	  if(ROW==7) Settings[S_MAINVOLTAGE_VBAT]=!Settings[S_MAINVOLTAGE_VBAT];
+	if(Screen.configPage == MENU_VOLTAGE && Screen.COL == 3) {
+	  if(Screen.ROW==1) Settings[S_DISPLAYVOLTAGE]=!Settings[S_DISPLAYVOLTAGE];
+	  if(Screen.ROW==2) Settings[S_DIVIDERRATIO]=Settings[S_DIVIDERRATIO]+_menudir;
+	  if(Screen.ROW==3) Settings[S_VOLTAGEMIN]=Settings[S_VOLTAGEMIN]+_menudir;
+	  if(Screen.ROW==4) Settings[S_VIDVOLTAGE]=!Settings[S_VIDVOLTAGE];
+	  if(Screen.ROW==5) Settings[S_VIDDIVIDERRATIO]=Settings[S_VIDDIVIDERRATIO]+_menudir;
+	  if(Screen.ROW==6) Settings[S_BATCELLS]=Settings[S_BATCELLS]+_menudir;
+	  if(Screen.ROW==7) Settings[S_MAINVOLTAGE_VBAT]=!Settings[S_MAINVOLTAGE_VBAT];
 	}
 #endif
 #ifdef MENU_RSSI
-	if(configPage == MENU_RSSI && COL == 3) {
-	  if(ROW==1) Settings[S_DISPLAYRSSI]=!Settings[S_DISPLAYRSSI];
-	  if(ROW==2) timer.rssiTimer=15; // 15 secs to turn off tx anwait to read min RSSI
-	  if(ROW==3) Settings[S_MWRSSI]=!Settings[S_MWRSSI];
-	  if(ROW==4) Settings[S_PWMRSSI]=!Settings[S_PWMRSSI];
-	  if(ROW==5) Settings16[S16_RSSIMAX]=Settings16[S16_RSSIMAX]+_menudir;
-	  if(ROW==6) Settings16[S16_RSSIMIN]=Settings16[S16_RSSIMIN]+_menudir;
+	if(Screen.configPage == MENU_RSSI && Screen.COL == 3) {
+	  if(Screen.ROW==1) Settings[S_DISPLAYRSSI]=!Settings[S_DISPLAYRSSI];
+	  if(Screen.ROW==2) timer.rssiTimer=15; // 15 secs to turn off tx anwait to read min RSSI
+	  if(Screen.ROW==3) Settings[S_MWRSSI]=!Settings[S_MWRSSI];
+	  if(Screen.ROW==4) Settings[S_PWMRSSI]=!Settings[S_PWMRSSI];
+	  if(Screen.ROW==5) Settings16[S16_RSSIMAX]=Settings16[S16_RSSIMAX]+_menudir;
+	  if(Screen.ROW==6) Settings16[S16_RSSIMIN]=Settings16[S16_RSSIMIN]+_menudir;
 	}
 #endif
 #ifdef MENU_CURRENT
-	if(configPage == MENU_CURRENT && COL == 3) {
-	  if(ROW==1) Settings[S_AMPERAGE]=!Settings[S_AMPERAGE];
-	  if(ROW==2) Settings[S_AMPER_HOUR]=!Settings[S_AMPER_HOUR];
-	  if(ROW==3) Settings[S_AMPERAGE_VIRTUAL]=!Settings[S_AMPERAGE_VIRTUAL];
-	  if(ROW==4) Settings16[S16_AMPDIVIDERRATIO]=Settings16[S16_AMPDIVIDERRATIO]+_menudir;
-	  if(ROW==5) Settings16[S16_AMPZERO]=Settings16[S16_AMPZERO]+_menudir;
+	if(Screen.configPage == MENU_CURRENT && Screen.COL == 3) {
+	  if(Screen.ROW==1) Settings[S_AMPERAGE]=!Settings[S_AMPERAGE];
+	  if(Screen.ROW==2) Settings[S_AMPER_HOUR]=!Settings[S_AMPER_HOUR];
+	  if(Screen.ROW==3) Settings[S_AMPERAGE_VIRTUAL]=!Settings[S_AMPERAGE_VIRTUAL];
+	  if(Screen.ROW==4) Settings16[S16_AMPDIVIDERRATIO]=Settings16[S16_AMPDIVIDERRATIO]+_menudir;
+	  if(Screen.ROW==5) Settings16[S16_AMPZERO]=Settings16[S16_AMPZERO]+_menudir;
 	}
 #endif
 #ifdef MENU_DISPLAY
-	if(configPage == MENU_DISPLAY && COL == 3) {
-	  if(ROW==1) Settings[S_DISPLAY_HORIZON_BR]=!Settings[S_DISPLAY_HORIZON_BR];
-	  if(ROW==2) Settings[S_WITHDECORATION]=!Settings[S_WITHDECORATION];
-	  if(ROW==3) Settings[S_SCROLLING]=!Settings[S_SCROLLING];
-	  if(ROW==4) Settings[S_THROTTLEPOSITION]=!Settings[S_THROTTLEPOSITION];
-	  if(ROW==5) Settings[S_COORDINATES]=!Settings[S_COORDINATES];
-	  if(ROW==6) Settings[S_MODESENSOR]=!Settings[S_MODESENSOR];
-	  if(ROW==7) Settings[S_GIMBAL]=!Settings[S_GIMBAL];
-	  if(ROW==8) Settings[S_MAPMODE]=Settings[S_MAPMODE]+_menudir;
+	if(Screen.configPage == MENU_DISPLAY && Screen.COL == 3) {
+	  if(Screen.ROW==1) Settings[S_DISPLAY_HORIZON_BR]=!Settings[S_DISPLAY_HORIZON_BR];
+	  if(Screen.ROW==2) Settings[S_WITHDECORATION]=!Settings[S_WITHDECORATION];
+	  if(Screen.ROW==3) Settings[S_SCROLLING]=!Settings[S_SCROLLING];
+	  if(Screen.ROW==4) Settings[S_THROTTLEPOSITION]=!Settings[S_THROTTLEPOSITION];
+	  if(Screen.ROW==5) Settings[S_COORDINATES]=!Settings[S_COORDINATES];
+	  if(Screen.ROW==6) Settings[S_MODESENSOR]=!Settings[S_MODESENSOR];
+	  if(Screen.ROW==7) Settings[S_GIMBAL]=!Settings[S_GIMBAL];
+	  if(Screen.ROW==8) Settings[S_MAPMODE]=Settings[S_MAPMODE]+_menudir;
 	}
 #endif
 #ifdef MENU_ADVANCED
-	if(configPage == MENU_ADVANCED && COL == 3) {
-	  if(ROW==1) Settings[S_UNITSYSTEM]=!Settings[S_UNITSYSTEM];
-	  if(ROW==2) {
+	if(Screen.configPage == MENU_ADVANCED && Screen.COL == 3) {
+	  if(Screen.ROW==1) Settings[S_UNITSYSTEM]=!Settings[S_UNITSYSTEM];
+	  if(Screen.ROW==2) {
 	    Settings[S_VIDEOSIGNALTYPE]=!Settings[S_VIDEOSIGNALTYPE];
-	    MAX7456.Setup();
+	    Max.Setup();
 	    }
-	  if(ROW==3) Settings[S_VREFERENCE]=!Settings[S_VREFERENCE];
-	  if(ROW==4) Settings[S_DEBUG]=!Settings[S_DEBUG];
-	  if(ROW==5) timer.magCalibrationTimer=CALIBRATION_DELAY;
-	  if(ROW==6) Settings[S_RCWSWITCH_CH]=Settings[S_RCWSWITCH_CH]+_menudir;	}
+	  if(Screen.ROW==3) Settings[S_VREFERENCE]=!Settings[S_VREFERENCE];
+	  if(Screen.ROW==4) Settings[S_DEBUG]=!Settings[S_DEBUG];
+	  if(Screen.ROW==5) timer.magCalibrationTimer=CALIBRATION_DELAY;
+	  if(Screen.ROW==6) Settings[S_RCWSWITCH_CH]=Settings[S_RCWSWITCH_CH]+_menudir;	}
 #endif
 #ifdef MENU_GPS_TIME
-	if(configPage == MENU_GPS_TIME && COL == 3) {
-	  if(ROW==1) Settings[S_GPSTIME]=!Settings[S_GPSTIME];
-	  if(ROW==2) Settings[S_GPSTZAHEAD]=!Settings[S_GPSTZAHEAD];
-	  if(ROW==3) if((_menudir == 1 && Settings[S_GPSTZ] < 130) || (_menudir == -1 && Settings[S_GPSTZ] > 0))Settings[S_GPSTZ]=Settings[S_GPSTZ]+_menudir*5;
+	if(Screen.configPage == MENU_GPS_TIME && Screen.COL == 3) {
+	  if(Screen.ROW==1) Settings[S_GPSTIME]=!Settings[S_GPSTIME];
+	  if(Screen.ROW==2) Settings[S_GPSTZAHEAD]=!Settings[S_GPSTZAHEAD];
+	  if(Screen.ROW==3) if((_menudir == 1 && Settings[S_GPSTZ] < 130) || (_menudir == -1 && Settings[S_GPSTZ] > 0))Settings[S_GPSTZ]=Settings[S_GPSTZ]+_menudir*5;
 	}
 #endif
 #ifdef MENU_ALARMS
-	if(configPage == MENU_ALARMS && COL == 3) {
-	  if(ROW==1) Settings[S_DISTANCE_ALARM]=Settings[S_DISTANCE_ALARM]+_menudir;
-	  if(ROW==2) Settings[S_ALTITUDE_ALARM]=Settings[S_ALTITUDE_ALARM]+_menudir;
-	  if(ROW==3) Settings[S_SPEED_ALARM]=Settings[S_SPEED_ALARM]+_menudir;
-	  if(ROW==4) Settings[S_FLYTIME_ALARM]=Settings[S_FLYTIME_ALARM]+_menudir;
-	  if(ROW==5) Settings[S_AMPER_HOUR_ALARM]=Settings[S_AMPER_HOUR_ALARM]+_menudir;
-	  if(ROW==6) Settings[S_AMPERAGE_ALARM]=Settings[S_AMPERAGE_ALARM]+_menudir;
+	if(Screen.configPage == MENU_ALARMS && Screen.COL == 3) {
+	  if(Screen.ROW==1) Settings[S_DISTANCE_ALARM]=Settings[S_DISTANCE_ALARM]+_menudir;
+	  if(Screen.ROW==2) Settings[S_ALTITUDE_ALARM]=Settings[S_ALTITUDE_ALARM]+_menudir;
+	  if(Screen.ROW==3) Settings[S_SPEED_ALARM]=Settings[S_SPEED_ALARM]+_menudir;
+	  if(Screen.ROW==4) Settings[S_FLYTIME_ALARM]=Settings[S_FLYTIME_ALARM]+_menudir;
+	  if(Screen.ROW==5) Settings[S_AMPER_HOUR_ALARM]=Settings[S_AMPER_HOUR_ALARM]+_menudir;
+	  if(Screen.ROW==6) Settings[S_AMPERAGE_ALARM]=Settings[S_AMPERAGE_ALARM]+_menudir;
 	}
 #endif
 #ifdef MENU_PROFILE
-	if(configPage == MENU_PROFILE && COL == 3) {
-	  if(ROW==1) FCProfile=FCProfile+_menudir;
-	  if(ROW==2) PIDController=PIDController+_menudir;
+	if(Screen.configPage == MENU_PROFILE && Screen.COL == 3) {
+	  if(Screen.ROW==1) FCProfile=FCProfile+_menudir;
+	  if(Screen.ROW==2) PIDController=PIDController+_menudir;
         #ifdef CORRECTLOOPTIME
-	  if(ROW==3) LoopTime=LoopTime+_menudir;
+	  if(Screen.ROW==3) LoopTime=LoopTime+_menudir;
         #endif
 	};
   #ifdef ENABLE_MSP_SAVE_ADVANCED
@@ -916,13 +916,13 @@ void MSPClass::serialMenuCommon()
         if (FCProfile!=PreviousFCProfile){
           setFCProfile();
           PreviousFCProfile=FCProfile;
-        }        
+        }
   #endif
-#endif  
-	if(ROW==10) {
-          previousconfigPage=configPage;
-	  if(COL==1) configExit();
-	  if(COL==2) configSave();
+#endif
+	if(Screen.ROW==10) {
+          Screen.previousconfigPage=Screen.configPage;
+	  if(Screen.COL==1) configExit();
+	  if(Screen.COL==2) configSave();
         }
 }
 
@@ -946,22 +946,22 @@ void MSPClass::handleRawRC() {
   if(!waitStick)
   {
     if((MwRcData[PITCHSTICK]>MAXSTICK)&&(MwRcData[YAWSTICK]>MAXSTICK)&&(MwRcData[THROTTLESTICK]>MINSTICK)){
-      if (!configMode&&(timer.allSec>5)&&!armed){
+      if (!configMode&&(timer.allSec>5)&&!mwosd.armed){
           // Enter config mode using stick combination
           waitStick =  2;	// Sticks must return to center before continue!
           configMode = 1;
-          configPage = previousconfigPage;
+          Screen.configPage = Screen.previousconfigPage;
           SetRequests();
       }
     }
     else if(configMode) {
       int8_t old_menudir=constrain(_menudir,-5,5);
       _menudir=0;
-      if(previousarmedstatus&&(MwRcData[THROTTLESTICK]>1300))
+      if(mwosd.previousarmedstatus&&(MwRcData[THROTTLESTICK]>1300))
       {
 	// EXIT from SHOW STATISTICS AFTER DISARM (push throttle up)
 	waitStick = 2;
-	ConfigExit();
+	configExit();
       }
 #ifdef MODE1
       if(configMode&&(MwRcData[YAWSTICK]>MAXSTICK)) // MOVE RIGHT
@@ -970,8 +970,8 @@ void MSPClass::handleRawRC() {
 #endif
       {
 	waitStick = 1;
-	COL++;
-	if(COL>3) COL=3;
+	Screen.COL++;
+	if(Screen.COL>3) Screen.COL=3;
       }
 #ifdef MODE1
       else if(configMode&&(MwRcData[YAWSTICK]<MINSTICK)) // MOVE LEFT
@@ -980,65 +980,75 @@ void MSPClass::handleRawRC() {
 #endif
       {
 	waitStick = 1;
-	COL--;
-	if(COL<1) COL=1;
+	Screen.COL--;
+	if(Screen.COL<1) Screen.COL=1;
       }
       else if(configMode&&(MwRcData[PITCHSTICK]>MAXSTICK)) // MOVE UP
       {
 	waitStick = 1;
-	ROW--;
-	if(ROW<1)
-	  ROW=1;
-        if(configPage == 0) {
-          ROW=10;
+	Screen.ROW--;
+	if(Screen.ROW<1)
+	  Screen.ROW=1;
+        if(Screen.configPage == 0) {
+          Screen.ROW=10;
         }
       }
       else if(configMode&&(MwRcData[PITCHSTICK]<MINSTICK)) // MOVE DOWN
       {
 	waitStick = 1;
-	ROW++;
-	if(ROW>10)
-	  ROW=10;
+	Screen.ROW++;
+	if(Screen.ROW>10)
+	  Screen.ROW=10;
       }
 #ifdef MODE1
-      else if(!previousarmedstatus&&configMode&&(MwRcData[ROLLSTICK]<MINSTICK)) // DECREASE
+      else if(!mwosd.previousarmedstatus&&configMode&&(MwRcData[ROLLSTICK]<MINSTICK)) // DECREASE
 #else
-      else if(!previousarmedstatus&&configMode&&(MwRcData[YAWSTICK]<MINSTICK)) // DECREASE
+      else if(!mwosd.previousarmedstatus&&configMode&&(MwRcData[YAWSTICK]<MINSTICK)) // DECREASE
 #endif
       {
 	waitStick = 1;
         _menudir=-1+old_menudir;
-        serialMenuCommon();  
+        serialMenuCommon();
       }
 #ifdef MODE1
-      else if(!previousarmedstatus&&configMode&&(MwRcData[ROLLSTICK]>MAXSTICK)) // INCREASE
+      else if(!mwosd.previousarmedstatus&&configMode&&(MwRcData[ROLLSTICK]>MAXSTICK)) // INCREASE
 #else
-      else if(!previousarmedstatus&&configMode&&(MwRcData[YAWSTICK]>MAXSTICK)) // INCREASE
+      else if(!mwosd.previousarmedstatus&&configMode&&(MwRcData[YAWSTICK]>MAXSTICK)) // INCREASE
 #endif
-      { 
+      {
 	waitStick =1;
         _menudir=1+old_menudir;
         #ifdef MENU_ALARMS
-	if(configPage == MENU_ALARMS && COL == 3) {
-	  if(ROW==5) timer.magCalibrationTimer=0;
+	if(Screen.configPage == MENU_ALARMS && Screen.COL == 3) {
+	  if(Screen.ROW==5) timer.magCalibrationTimer=0;
         }
         #endif //MENU_ALARMS
-        serialMenuCommon();  
-      }      
+        serialMenuCommon();
+      }
     }
     if(waitStick == 1)
       stickTime = millis();
   }
 }
 
-void MSPClass::ConfigExit()
+void MSPClass::ConfigEnter()
 {
-  configPage=1;
-  ROW=10;
-  COL=3;
+  timer.armed=20;
+  Screen.configPage=0;
+  Screen.ROW=10;
+  Screen.COL=1;
+  configMode=1;
+  SetRequests();
+}
+
+void MSPClass::configExit()
+{
+  Screen.configPage=1;
+  Screen.ROW=10;
+  Screen.COL=3;
   configMode=0;
   //waitStick=3;
-  previousarmedstatus = 0;
+  mwosd.previousarmedstatus = 0;
   if (Settings[S_RESETSTATISTICS]){
     Stats.Reset();
   }
@@ -1062,60 +1072,60 @@ void MSPClass::configSave()
 
   WriteRequest(MSP_SET_LOOP_TIME, 2);
   write16(LoopTime);
-  writeChecksum();  
+  writeChecksum();
 #endif
 
   WriteRequest(MSP_SET_PID, PIDITEMS*3);
   for(uint8_t i=0; i<PIDITEMS; i++) {
-    write8(P8[i]);
-    write8(I8[i]);
-    write8(D8[i]);
+    write8(mwosd.P8[i]);
+    write8(mwosd.I8[i]);
+    write8(mwosd.D8[i]);
   }
   writeChecksum();
-  
+
 #if defined CORRECT_MSP_CF2
   WriteRequest(MSP_SET_RC_TUNING,11);
-  write8(rcRate8);
-  write8(rcExpo8);
-  write8(rollRate);
-  write8(PitchRate);
-  write8(yawRate);
-  write8(dynThrPID);
-  write8(thrMid8);
-  write8(thrExpo8);
-  write16(tpa_breakpoint16);
-  write8(rcYawExpo8);
+  write8(rc.rcRate8);
+  write8(rc.rcExpo8);
+  write8(rc.rollRate);
+  write8(rc.PitchRate);
+  write8(rc.yawRate);
+  write8(rc.dynThrPID);
+  write8(rc.thrMid8);
+  write8(rc.thrExpo8);
+  write16(rc.tpa_breakpoint16);
+  write8(rc.rcYawExpo8);
   writeChecksum();
 #elif defined CORRECT_MSP_CF1
   WriteRequest(MSP_SET_RC_TUNING,10);
-  write8(rcRate8);
-  write8(rcExpo8);
-  write8(rollRate);
-  write8(PitchRate);
-  write8(yawRate);
-  write8(dynThrPID);
-  write8(thrMid8);
-  write8(thrExpo8);
-  write16(tpa_breakpoint16);
+  write8(rc.rcRate8);
+  write8(rc.rcExpo8);
+  write8(rc.rollRate);
+  write8(rc.PitchRate);
+  write8(rc.yawRate);
+  write8(rc.dynThrPID);
+  write8(rc.thrMid8);
+  write8(rc.thrExpo8);
+  write16(rc.tpa_breakpoint16);
   writeChecksum();
 #else
   WriteRequest(MSP_SET_RC_TUNING,7);
-  write8(rcRate8);
-  write8(rcExpo8);
-  write8(rollPitchRate);
-  write8(yawRate);
-  write8(dynThrPID);
-  write8(thrMid8);
-  write8(thrExpo8);
+  write8(rc.rcRate8);
+  write8(rc.rcExpo8);
+  write8(rc.rollPitchRate);
+  write8(rc.yawRate);
+  write8(rc.dynThrPID);
+  write8(rc.thrMid8);
+  write8(rc.thrExpo8);
   writeChecksum();
  #endif
 
 #if defined CORRECT_MSP_BF1
   WriteRequest(MSP_SET_CONFIG,25);
-  bfconfig[18] =rollRate;
-  bfconfig[19] =PitchRate;
+  _bfconfig[18] =rollRate;
+  _bfconfig[19] =PitchRate;
   for(uint8_t i=0; i<25; i++) {
-    write8(bfconfig[i]);
+    write8(_bfconfig[i]);
   }
   writeChecksum();
 #endif
@@ -1162,4 +1172,21 @@ void MSPClass::setFCProfile()
   WriteRequest(MSP_EEPROM_WRITE, 0);
   SetRequests();
   delay(100);
+}
+
+void MSPClass::Countdown()
+{
+  if (timer.MSP_active>0){
+    timer.MSP_active--;
+  }
+}
+
+void MSPClass::ConfigExitIfArmed()
+{
+  if(mwosd.armed){
+    mwosd.previousarmedstatus=1;
+    if (configMode==1) {
+      configExit();
+    }
+  }
 }
